requirements.lock.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.52.0
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
cryptography==42.0.8
debugpy==1.8.13
decorator==5.2.1
distro==1.9.0
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.17.0
flake8==7.2.0
Flask==3.1.0
flatbuffers==25.1.24
flwr==1.14.0
frozenlist==1.6.0
fsspec==2024.12.0
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h5py==3.12.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
iterators==0.0.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.8.0
langchain==0.3.25
langchain-core==0.3.61
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.1
langgraph-sdk==0.1.70
langsmith==0.3.42
libclang==18.1.1
littleutils==0.2.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
ogb==1.3.6
openai==1.82.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.18
ormsgpack==1.10.0
outdated==0.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.6.0
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.11.5
pydantic_core==2.33.2
pyee==11.1.1
pyflakes==3.3.2
Pygments==2.19.1
pyparsing==3.2.3
pyppeteer==2.0.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data==0.6.3
sympy==1.13.3
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==2.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tomli==2.2.1
tomli_w==1.2.0
torch==2.2.2
torch-geometric==2.6.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.12.5
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
wcwidth==0.2.13
websockets==10.4
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
